refactor: renomme le package glose en sigma
renomme Gloses -> StraightPos2Sigmas
renomme CGloses -> ConstrainedPos2Sigmas
refactor __call__ dans ConstrainedPos2Sigmas
