+++ pfmg/lexique/sigma/Sigma.py
