+++ pfmg/lexique/sigma/test/__init__.py
